requirements.txt
# AI Control Tower for U.S. Supply-Chain Resilience (v0.0.1)
# Core dependencies for type-safe risk signal processing and resilience reporting

# Pydantic for type-safe data validation and models
pydantic>=2.0

# PydanticAI for AI agent orchestration (Claude 3.5 Sonnet backend)
# Version 0.0.13+ uses model strings (e.g., 'anthropic:claude-3-5-sonnet-latest')
pydantic-ai>=0.0.13

# Logfire for observability and structured logging
logfire

# python-dotenv for environment variable management
python-dotenv

# LangChain core + community integrations for RAG (still available but no longer required for core TF-IDF RAG)
langchain>=0.2
langchain-community>=0.2

# PDF parsing for legislative documents
pypdf
llama-parse

# Scikit-learn and joblib for local TF-IDF vectorization (no Torch/GPU required)
scikit-learn
joblib

# Geospatial radius filtering + map visualization
geopy
folium
requests
pyyaml
gradio
langgraph

# System dependency (not a pip package):
# Install Ollama from https://ollama.com/download